Entry Now Works
Hardcoding the dependant column / y column is no longer needed. I fixed the entry field.
I changed up the logic. I forgot that tkinter passes all entry fields as strings, but I was creating the logic that somehow it wasn't, that like 2 would be passed as 2, not "2". So I cleaned up and massively simplified the entry field logic to just use .isDigit() against the entry value.
✅Proper Entry Field
🚫No entry field error handing.. i.e. "That column doesn't exist, or 25 is outside the range of the columns in the csv"
🚫All previous "🚫" fields in the previous commit.